+++ examples/example-react-sdk-2.x/src/components/NextSteps.tsx
@@ -5,7 +5,7 @@ import css from "./NextSteps.module.css";
 
 const STEPS: INextStep[] = [
   {
-    icon: "tag",
+    icon: "rocket-slant",
     title: "Deploy your application",
     subtitle: (
       <span>