+++ packages/create-app.template.react.beta/templates/src/components/Header.tsx
@@ -1,13 +1,13 @@
 import palantir from "/palantir.svg";
 import react from "/react.svg";
 
-import { H5, Link } from "@blueprintjs/core";
+import { H2, Link } from "@blueprintjs/core";
 import React from "react";
 import css from "./Header.module.css";
 
 function Header(): React.ReactElement {
   return (
-    <div className={css.headerContainer}>
+    <header className={css.headerContainer}>
       <div className={css.logos}>
         <Link
           href="https://www.palantir.com/docs/foundry/ontology-sdk/overview/"
@@ -20,8 +20,8 @@ function Header(): React.ReactElement {
           <img src={react} className={css.logo} alt="React logo" />
         </Link>
       </div>
-      <H5 className={css.header}>Build your React application with Foundry!</H5>
-    </div>
+      <H2 className={css.header}>Build your React application with Foundry!</H2>
+    </header>
   );
 }