+++ packages/react-components/src/filter-list/base/FilterList.tsx
@@ -39,6 +39,7 @@ export function FilterList<Q extends ObjectTypeDefinition>(
     onReset,
     onFilterAdded,
     showActiveFilterCount = false,
+    enableSorting,
     className,
     renderAddFilterButton,
   } = props;
@@ -93,6 +94,7 @@ export function FilterList<Q extends ObjectTypeDefinition>(
         filterStates={filterStates}
         onFilterStateChanged={setFilterState}
         whereClause={whereClause}
+        enableSorting={enableSorting}
       />
 
       {showAddButton && (