+++ packages/react/src/public/experimental.ts
@@ -15,9 +15,6 @@
  */
 
 export { OsdkProvider2 } from "../new/OsdkProvider2.js";
-export { useCurrentFoundryUser } from "../new/platform-apis/admin/useCurrentFoundryUser.js";
-export { useFoundryUser } from "../new/platform-apis/admin/useFoundryUser.js";
-export { useFoundryUsersList } from "../new/platform-apis/admin/useFoundryUsersList.js";
 export { useLinks } from "../new/useLinks.js";
 export { useObjectSet } from "../new/useObjectSet.js";
 export { useOsdkAction } from "../new/useOsdkAction.js";