+++ packages/create-app.template.vue.v2/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @osdk/create-app.template.vue.v2
 
+## 2.8.0-beta.16
+
+### Minor Changes
+
+- 0ff602a: Bump vite to ^7.3.1
+- e22f91c: Make template a11y compatible
+
 ## 2.8.0-beta.15
 
 ## 2.8.0-beta.14