+++ packages/create-widget.template.react.v2/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/create-widget.template.react.v2
 
+## 3.5.0-beta.16
+
+### Minor Changes
+
+- 0ff602a: Bump vite to ^7.3.1
+
 ## 3.5.0-beta.15
 
 ## 3.5.0-beta.14